Upgrade to EhCache 3.5.0
Closes gh-12256
