+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/cache/CacheAutoConfigurationTests.java
@@ -16,11 +16,7 @@
 
 package org.springframework.boot.autoconfigure.cache;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
-import java.util.Map;
 
 import javax.cache.Caching;
 import javax.cache.configuration.CompleteConfiguration;
@@ -39,7 +35,6 @@ import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.spring.cache.HazelcastCacheManager;
 import net.sf.ehcache.Status;
-import org.ehcache.jsr107.EhcacheCachingProvider;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.jcache.embedded.JCachingProvider;
 import org.infinispan.spring.provider.SpringEmbeddedCacheManager;
@@ -52,8 +47,6 @@ import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.autoconfigure.cache.support.MockCachingProvider;
 import org.springframework.boot.autoconfigure.hazelcast.HazelcastAutoConfiguration;
 import org.springframework.boot.test.context.assertj.AssertableApplicationContext;
-import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-import org.springframework.boot.test.context.runner.ContextConsumer;
 import org.springframework.boot.testsupport.runner.classpath.ClassPathExclusions;
 import org.springframework.boot.testsupport.runner.classpath.ModifiedClassPathRunner;
 import org.springframework.cache.Cache;
@@ -64,7 +57,6 @@ import org.springframework.cache.caffeine.CaffeineCache;
 import org.springframework.cache.caffeine.CaffeineCacheManager;
 import org.springframework.cache.concurrent.ConcurrentMapCache;
 import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
-import org.springframework.cache.ehcache.EhCacheCacheManager;
 import org.springframework.cache.interceptor.CacheResolver;
 import org.springframework.cache.jcache.JCacheCacheManager;
 import org.springframework.cache.support.NoOpCacheManager;
@@ -93,10 +85,7 @@ import static org.mockito.Mockito.verify;
  */
 @RunWith(ModifiedClassPathRunner.class)
 @ClassPathExclusions("hazelcast-client-*.jar")
-public class CacheAutoConfigurationTests {
-
-	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
-			.withConfiguration(AutoConfigurations.of(CacheAutoConfiguration.class));
+public class CacheAutoConfigurationTests extends AbstractCacheAutoConfigurationTests {
 
 	@Test
 	public void noEnableCaching() {
@@ -161,7 +150,8 @@ public class CacheAutoConfigurationTests {
 		this.contextRunner
 				.withUserConfiguration(DefaultCacheAndCustomizersConfiguration.class)
 				.withPropertyValues("spring.cache.type=" + "simple")
-				.run(dunno("allCacheManagerCustomizer", "simpleCacheManagerCustomizer"));
+				.run(verifyCustomizers("allCacheManagerCustomizer",
+						"simpleCacheManagerCustomizer"));
 	}
 
 	@Test
@@ -206,7 +196,8 @@ public class CacheAutoConfigurationTests {
 		this.contextRunner
 				.withUserConfiguration(GenericCacheAndCustomizersConfiguration.class)
 				.withPropertyValues("spring.cache.type=" + "generic")
-				.run(dunno("allCacheManagerCustomizer", "genericCacheManagerCustomizer"));
+				.run(verifyCustomizers("allCacheManagerCustomizer",
+						"genericCacheManagerCustomizer"));
 	}
 
 	@Test
@@ -237,8 +228,9 @@ public class CacheAutoConfigurationTests {
 	public void couchbaseCacheWithCustomizers() {
 		this.contextRunner
 				.withUserConfiguration(CouchbaseCacheAndCustomizersConfiguration.class)
-				.withPropertyValues("spring.cache.type=" + "couchbase").run(dunno(
-						"allCacheManagerCustomizer", "couchbaseCacheManagerCustomizer"));
+				.withPropertyValues("spring.cache.type=" + "couchbase")
+				.run(verifyCustomizers("allCacheManagerCustomizer",
+						"couchbaseCacheManagerCustomizer"));
 	}
 
 	@Test
@@ -324,8 +316,8 @@ public class CacheAutoConfigurationTests {
 	@Test
 	public void redisCacheWithCustomizers() {
 		this.contextRunner.withUserConfiguration(RedisWithCustomizersConfiguration.class)
-				.withPropertyValues("spring.cache.type=" + "redis")
-				.run(dunno("allCacheManagerCustomizer", "redisCacheManagerCustomizer"));
+				.withPropertyValues("spring.cache.type=" + "redis").run(verifyCustomizers(
+						"allCacheManagerCustomizer", "redisCacheManagerCustomizer"));
 	}
 
 	@Test
@@ -475,85 +467,6 @@ public class CacheAutoConfigurationTests {
 	}
 
 	@Test
-	public void ehcacheCacheWithCaches() {
-		this.contextRunner.withUserConfiguration(DefaultCacheConfiguration.class)
-				.withPropertyValues("spring.cache.type=ehcache").run((context) -> {
-					EhCacheCacheManager cacheManager = getCacheManager(context,
-							EhCacheCacheManager.class);
-					assertThat(cacheManager.getCacheNames()).containsOnly("cacheTest1",
-							"cacheTest2");
-					assertThat(context.getBean(net.sf.ehcache.CacheManager.class))
-							.isEqualTo(cacheManager.getCacheManager());
-				});
-	}
-
-	@Test
-	public void ehcacheCacheWithCustomizers() {
-		this.contextRunner
-				.withUserConfiguration(DefaultCacheAndCustomizersConfiguration.class)
-				.withPropertyValues("spring.cache.type=" + "ehcache")
-				.run(dunno("allCacheManagerCustomizer", "ehcacheCacheManagerCustomizer"));
-	}
-
-	@Test
-	public void ehcacheCacheWithConfig() {
-		this.contextRunner.withUserConfiguration(DefaultCacheConfiguration.class)
-				.withPropertyValues("spring.cache.type=ehcache",
-						"spring.cache.ehcache.config=cache/ehcache-override.xml")
-				.run((context) -> {
-					EhCacheCacheManager cacheManager = getCacheManager(context,
-							EhCacheCacheManager.class);
-					assertThat(cacheManager.getCacheNames())
-							.containsOnly("cacheOverrideTest1", "cacheOverrideTest2");
-				});
-	}
-
-	@Test
-	public void ehcacheCacheWithExistingCacheManager() {
-		this.contextRunner.withUserConfiguration(EhCacheCustomCacheManager.class)
-				.withPropertyValues("spring.cache.type=ehcache").run((context) -> {
-					EhCacheCacheManager cacheManager = getCacheManager(context,
-							EhCacheCacheManager.class);
-					assertThat(cacheManager.getCacheManager())
-							.isEqualTo(context.getBean("customEhCacheCacheManager"));
-				});
-	}
-
-	@Test
-	public void ehcache3AsJCacheWithCaches() {
-		String cachingProviderFqn = EhcacheCachingProvider.class.getName();
-		this.contextRunner.withUserConfiguration(DefaultCacheConfiguration.class)
-				.withPropertyValues("spring.cache.type=jcache",
-						"spring.cache.jcache.provider=" + cachingProviderFqn,
-						"spring.cache.cacheNames[0]=foo",
-						"spring.cache.cacheNames[1]=bar")
-				.run((context) -> {
-					JCacheCacheManager cacheManager = getCacheManager(context,
-							JCacheCacheManager.class);
-					assertThat(cacheManager.getCacheNames()).containsOnly("foo", "bar");
-				});
-	}
-
-	@Test
-	public void ehcache3AsJCacheWithConfig() {
-		String cachingProviderFqn = EhcacheCachingProvider.class.getName();
-		String configLocation = "ehcache3.xml";
-		this.contextRunner.withUserConfiguration(DefaultCacheConfiguration.class)
-				.withPropertyValues("spring.cache.type=jcache",
-						"spring.cache.jcache.provider=" + cachingProviderFqn,
-						"spring.cache.jcache.config=" + configLocation)
-				.run((context) -> {
-					JCacheCacheManager cacheManager = getCacheManager(context,
-							JCacheCacheManager.class);
-
-					Resource configResource = new ClassPathResource(configLocation);
-					assertThat(cacheManager.getCacheManager().getURI())
-							.isEqualTo(configResource.getURI());
-					assertThat(cacheManager.getCacheNames()).containsOnly("foo", "bar");
-				});
-	}
-
-	@Test
 	public void hazelcastCacheExplicit() {
 		this.contextRunner
 				.withConfiguration(
@@ -575,8 +488,9 @@ public class CacheAutoConfigurationTests {
 	public void hazelcastCacheWithCustomizers() {
 		this.contextRunner
 				.withUserConfiguration(HazelcastCacheAndCustomizersConfiguration.class)
-				.withPropertyValues("spring.cache.type=" + "hazelcast").run(dunno(
-						"allCacheManagerCustomizer", "hazelcastCacheManagerCustomizer"));
+				.withPropertyValues("spring.cache.type=" + "hazelcast")
+				.run(verifyCustomizers("allCacheManagerCustomizer",
+						"hazelcastCacheManagerCustomizer"));
 	}
 
 	@Test
@@ -700,8 +614,9 @@ public class CacheAutoConfigurationTests {
 	public void infinispanCacheWithCustomizers() {
 		this.contextRunner
 				.withUserConfiguration(DefaultCacheAndCustomizersConfiguration.class)
-				.withPropertyValues("spring.cache.type=" + "infinispan").run(dunno(
-						"allCacheManagerCustomizer", "infinispanCacheManagerCustomizer"));
+				.withPropertyValues("spring.cache.type=" + "infinispan")
+				.run(verifyCustomizers("allCacheManagerCustomizer",
+						"infinispanCacheManagerCustomizer"));
 	}
 
 	@Test
@@ -798,8 +713,9 @@ public class CacheAutoConfigurationTests {
 	public void caffeineCacheWithCustomizers() {
 		this.contextRunner
 				.withUserConfiguration(DefaultCacheAndCustomizersConfiguration.class)
-				.withPropertyValues("spring.cache.type=" + "caffeine").run(dunno(
-						"allCacheManagerCustomizer", "caffeineCacheManagerCustomizer"));
+				.withPropertyValues("spring.cache.type=" + "caffeine")
+				.run(verifyCustomizers("allCacheManagerCustomizer",
+						"caffeineCacheManagerCustomizer"));
 	}
 
 	@Test
@@ -837,35 +753,6 @@ public class CacheAutoConfigurationTests {
 		foo.get("1");
 		assertThat(((CaffeineCache) foo).getNativeCache().stats().missCount())
 				.isEqualTo(1L);
-	}
-
-	@SuppressWarnings("rawtypes")
-	private ContextConsumer<AssertableApplicationContext> dunno(
-			String... expectedCustomizerNames) {
-		return (context) -> {
-			CacheManager cacheManager = getCacheManager(context, CacheManager.class);
-			List<String> expected = new ArrayList<>(
-					Arrays.asList(expectedCustomizerNames));
-			Map<String, CacheManagerTestCustomizer> customizer = context
-					.getBeansOfType(CacheManagerTestCustomizer.class);
-			customizer.forEach((key, value) -> {
-				if (expected.contains(key)) {
-					expected.remove(key);
-					assertThat(value.cacheManager).isSameAs(cacheManager);
-				}
-				else {
-					assertThat(value.cacheManager).isNull();
-				}
-			});
-			assertThat(expected).hasSize(0);
-		};
-	}
-
-	private <T extends CacheManager> T getCacheManager(
-			AssertableApplicationContext loaded, Class<T> type) {
-		CacheManager cacheManager = loaded.getBean(CacheManager.class);
-		assertThat(cacheManager).as("Wrong cache manager type").isInstanceOf(type);
-		return type.cast(cacheManager);
 	}
 
 	@Configuration
@@ -1118,89 +1005,6 @@ public class CacheAutoConfigurationTests {
 		@Bean
 		CaffeineSpec caffeineSpec() {
 			return CaffeineSpec.parse("recordStats");
-		}
-
-	}
-
-	@Configuration
-	static class CacheManagerCustomizersConfiguration {
-
-		@Bean
-		public CacheManagerCustomizer<CacheManager> allCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<CacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<ConcurrentMapCacheManager> simpleCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<ConcurrentMapCacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<SimpleCacheManager> genericCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<SimpleCacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<CouchbaseCacheManager> couchbaseCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<CouchbaseCacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<RedisCacheManager> redisCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<RedisCacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<EhCacheCacheManager> ehcacheCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<EhCacheCacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<HazelcastCacheManager> hazelcastCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<HazelcastCacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<SpringEmbeddedCacheManager> infinispanCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<SpringEmbeddedCacheManager>() {
-
-			};
-		}
-
-		@Bean
-		public CacheManagerCustomizer<CaffeineCacheManager> caffeineCacheManagerCustomizer() {
-			return new CacheManagerTestCustomizer<CaffeineCacheManager>() {
-
-			};
-		}
-
-	}
-
-	static abstract class CacheManagerTestCustomizer<T extends CacheManager>
-			implements CacheManagerCustomizer<T> {
-
-		private T cacheManager;
-
-		@Override
-		public void customize(T cacheManager) {
-			if (this.cacheManager != null) {
-				throw new IllegalStateException("Customized invoked twice");
-			}
-			this.cacheManager = cacheManager;
 		}
 
 	}
