+++ spring-boot-project/spring-boot-dependencies/pom.xml
@@ -47,7 +47,7 @@
 		<dom4j.version>1.6.1</dom4j.version>
 		<dropwizard-metrics.version>3.2.6</dropwizard-metrics.version>
 		<ehcache.version>2.10.4</ehcache.version>
-		<ehcache3.version>3.4.0</ehcache3.version>
+		<ehcache3.version>3.5.0</ehcache3.version>
 		<embedded-mongo.version>2.0.3</embedded-mongo.version>
 		<flyway.version>5.0.7</flyway.version>
 		<freemarker.version>2.3.27-incubating</freemarker.version>
